restructured files and added a few more routes
